refactor: Move function to utils
