+++ app/components/chart-preview.tsx
@@ -6,7 +6,6 @@ import {
   useDraggable,
   useDroppable,
 } from "@dnd-kit/core";
-import { getEventCoordinates } from "@dnd-kit/utilities";
 import { Trans } from "@lingui/macro";
 import { Box, useMediaQuery } from "@mui/material";
 import Head from "next/head";
@@ -52,9 +51,8 @@ import { useLocale } from "@/locales/use-locale";
 import { InteractiveFiltersProvider } from "@/stores/interactive-filters";
 import { useTransitionStore } from "@/stores/transition";
 import { useTheme } from "@/themes";
+import { snapCornerToCursor } from "@/utils/dnd";
 import useEvent from "@/utils/use-event";
-
-import type { Modifier } from "@dnd-kit/core";
 
 type TallLayoutRowProps = {
   row: TallLayoutRow;
@@ -620,28 +618,4 @@ export const ChartPreviewInner = (props: ChartPreviewInnerProps) => {
       </ChartErrorBoundary>
     </Flex>
   );
-};
-
-const snapCornerToCursor: Modifier = ({
-  activatorEvent,
-  draggingNodeRect,
-  transform,
-}) => {
-  if (draggingNodeRect && activatorEvent) {
-    const activatorCoordinates = getEventCoordinates(activatorEvent);
-
-    if (!activatorCoordinates) {
-      return transform;
-    }
-
-    const offsetX = activatorCoordinates.x - draggingNodeRect.left + 48;
-
-    return {
-      ...transform,
-      x: transform.x + offsetX - draggingNodeRect.width,
-      y: transform.y,
-    };
-  }
-
-  return transform;
 };
