+++ app/utils/dnd.ts
@@ -1,0 +1,26 @@
+import { Modifier } from "@dnd-kit/core";
+import { getEventCoordinates } from "@dnd-kit/utilities";
+
+export const snapCornerToCursor: Modifier = ({
+  activatorEvent,
+  draggingNodeRect,
+  transform,
+}) => {
+  if (draggingNodeRect && activatorEvent) {
+    const activatorCoordinates = getEventCoordinates(activatorEvent);
+
+    if (!activatorCoordinates) {
+      return transform;
+    }
+
+    const offsetX = activatorCoordinates.x - draggingNodeRect.left + 48;
+
+    return {
+      ...transform,
+      x: transform.x + offsetX - draggingNodeRect.width,
+      y: transform.y,
+    };
+  }
+
+  return transform;
+};
